Refactor httpLogging to requestLogging
* refactor httpLogging > requestLogging: more clear naming
* deps
  * prod
    * change fastify: 4.16.0 > 4.16.3
