+++ package.json
@@ -106,7 +106,7 @@
     "error-stack-parser": "^2.1.4",
     "fast-json-stringify": "^5.7.0",
     "fast-safe-stringify": "^2.1.1",
-    "fastify": "^4.16.0",
+    "fastify": "^4.16.3",
     "fastify-plugin": "^4.5.0",
     "http-status-codes": "^2.2.0",
     "jin-curlize": "^0.5.0",
