+++ src/loggers/modules/getHttpMethod.ts
